+++ android/src/main/java/com/rnoriginalscrollview/ReactScrollView.java
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-package com.rnnestedscrollview;
+package com.rnoriginalscrollview;
 
 import android.graphics.Canvas;
 import android.graphics.Color;
