+++ index.js
@@ -1,5 +1,5 @@
 import { Platform, ScrollView } from 'react-native';
 
 export default Platform.OS === 'android'
-  ? require('./lib/NestedScrollView')
+  ? require('./lib/OriginalScrollView')
   : ScrollView;
